move login + auth to /auth/
