--- components/views/login/Login.module.scss
@@ -1,81 +1,0 @@
-.content {
-    display: grid;
-    place-items: center;
-    height: calc(100vh - var(--nav-height) - var(--footer-height));
-    .formContainer {
-        height: fit-content;
-        width: 20rem;
-        background: var(--darkClearBg);
-        padding: 2rem;
-        display: flex;
-        flex-direction: column;
-        .header {
-            text-align: center;
-            color: var(--off-white);
-            padding-bottom: 1rem;
-        }
-        .errorMessage {
-            color: var(--error-red);
-        }
-        .formInputs {
-            display: flex;
-            flex-direction: column;
-            label {
-                display: flex;
-                flex-direction: column;
-                align-items: flex-start;
-                padding: 0.2rem 0;
-                color: var(--off-white);
-            }
-            input {
-                position: relative;
-                margin-bottom: 1rem;
-                width: 100%;
-                display: block;
-                font-size: 1rem;
-
-                padding: 8px;
-                border-radius: 6px;
-                border: none;
-                background: rgba(3, 3, 3, 0.1);
-                transition: all 0.2s ease-in-out;
-                &::placeholder {
-                    font-size: 12px;
-                    color: rgba(0, 0, 0, 0.364);
-                    font-weight: 600;
-                }
-                &:focus {
-                    outline: none;
-                    box-shadow: 3px 3px 10px #333;
-                    background: rgba(3, 3, 3, 0.18);
-                }
-            }
-
-            .button {
-                cursor: pointer;
-                text-decoration: none;
-                position: relative;
-                width: 100%;
-                display: block;
-                margin: 9px auto;
-                font-size: 17px;
-                color: var(--off-white);
-                padding: 8px;
-                border-radius: 6px;
-                border: none;
-                background: rgba(3, 3, 3, 0.1);
-                transition: all 0.2s ease-in-out;
-                &:hover {
-                    background: rgba(3, 3, 3, 0.648);
-                }
-            }
-            .forgotPw {
-                font-size: 11px;
-                text-align: center;
-                color: var(--off-white);
-                padding: 0.5rem 0;
-                text-decoration: underline;
-            }
-        }
-    }
-}
